Implement the previously discussed changes
When registering an array now, you pass in the number of elements and a name as description.

This gets stored for each passed in array.

Later the index can be reconstructed by taking a look at the current memory address and the size of a single element which should be in that memory region.