Add function to generate valid JSON
This also made the generators public static methods, so they can be used. 
I had to correct a previous mistake, that the memory regions also need to be passed into the the function so it doesn't depend on a local existing object (as then you couldn't pass in own functions).