Change the opacity for the accesses (heatmap), toggle background, split
With this commit the opacity is automatically changed to fit the maximum access and the current access.

Additionally two toggle buttons were introduced, one to toggle either black or white background for the memory table, the other to toggle either the unified memory view or the split (into read and write) memory view.