Improve visibility of elements
Added a toggle to not show the current memory index, allowing to fit memory more densely. (You can still see the index by hovering for a medium long time).

To speed up page rendering, background color has been moved to outside of the single component, so that just the parent needs to be updated.

Additionally to be able to see the different cells more easily, borders have been added.

By default for better visibility elements will get wrapped to show as many on screen as possible, to easier identify hotspots.