Annotate all examples, function to generate just JSON, small improvments
For every example there is now an annotated version (which is also in the makefile) and the original version of that code. (It is mostly sufficient to change about 12 lines of code).

The cuda_mav class has been improved, by being able to directly produce JSON files and strings as output.
The currentSize and originalSize have been added to the output, so you know if it was not possible to store all data which was actually accessed. Instead of a tuple, a dedicated struct is now used for memory regions. 

Templates have been moved to their own dedicated directory.

To reduce git load times a correct gitignore was already added to the html folder. This will be solved later by the merge into master anyways, but eases local development.