Add information about the current memory region to the sidebar
This information contains both the grid and block size, as well as the warp size, but also information about the memory region like start and end address, amount of elements and the total count of accesses in this memory region.

Also in the case that the memory storage did not allocate enough elements to store all accesses, display a warning in the file.