Fix incorrect block, thread and grid mixups
(Main mistake was that the order of parameters in AccessInstance was wrong, which lead me down a debugging session.

Also for better naming changed global to local.
As the respective ids are not the entirely global ids, but for example the local thread id in the current block, or the local block id in the current grid